+++ client/src/components/Home.js
@@ -11,10 +11,10 @@ const Home = () => {
   // INSERT YOUR CREATED MOVIE ENDPOINT
 
   const MOVIES_VECTOR_SEARCH_ENDPOINT =
-    "http://localhost:5050/getSemanticMovieSearch";
+    "http://localhost:5050/movies/semantic";
 
   let MOVIES_VECTOR_SEARCH_ADVANCED_ENDPOINT =
-    "http://localhost:5050/searchMoviesAdvanced";
+    "http://localhost:5050/movies/semantic/advanced";
 
   const [movies, setMovies] = useState([]);
   const [searchTerm, setSearchTerm] = useState("");
